fix-refactor: google 로그인 이슈 해결 및 프로젝트 리펙토링
