+++ android/app/build.gradle.kts
@@ -6,7 +6,7 @@ plugins {
 }
 
 android {
-    namespace = "com.example.untitled4"
+    namespace = "com.nam9.lami"
     compileSdk = flutter.compileSdkVersion
     // NDK 버전을 최신 버전으로 업데이트
     ndkVersion = "27.0.12077973"
@@ -23,7 +23,7 @@ android {
 
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId = "com.example.untitled4"
+        applicationId = "com.nam9.lami"
         // You can update the following values to match your application needs.
         // For more information, see: https://flutter.dev/to/review-gradle-config.
         minSdk = flutter.minSdkVersion
