+++ android/app/src/main/kotlin/com/nam9/lami/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.untitled4
+package com.nam9.lami
 
 import io.flutter.embedding.android.FlutterActivity
