+++ lib/favorite_service.dart
@@ -1,9 +1,6 @@
-import 'dart:convert';
-import 'dart:io' show Platform;
 import 'package:flutter/foundation.dart';
-import 'package:http/http.dart' as http;
-import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'services/favorite_api_service.dart';
 import 'models/favorite_route_model.dart';
 
 class FavoriteService extends ChangeNotifier {
@@ -14,25 +11,10 @@ class FavoriteService extends ChangeNotifier {
     final prefs = await SharedPreferences.getInstance();
     final googleId = prefs.getString('googleId');
     if (googleId == null) return;
-    // Determine base URL with platform fallback
-    final defaultUrl = dotenv.env['BACKEND_URL'] ?? 'http://localhost:3000';
-    final baseUrl = Platform.isAndroid
-        ? (dotenv.env['SERVER_BASE_URL_ANDROID'] ?? defaultUrl)
-        : Platform.isIOS
-            ? (dotenv.env['SERVER_BASE_URL_IOS'] ?? defaultUrl)
-            : defaultUrl;
-    final uri = Uri.parse('$baseUrl/traffic/routes/favorites?googleId=$googleId');
     try {
-      final resp = await http.get(uri);
-      if (resp.statusCode >= 200 && resp.statusCode < 300) {
-        final List<dynamic> data = json.decode(resp.body);
-        _favoriteList = data
-            .map((e) => FavoriteRouteModel.fromJson(e as Map<String, dynamic>))
-            .toList();
-        notifyListeners();
-      } else {
-        debugPrint('즐겨찾기 불러오기 실패: ${resp.body}');
-      }
+      final list = await FavoriteApiService(googleId: googleId).getFavorites();
+      _favoriteList = list;
+      notifyListeners();
     } catch (e) {
       debugPrint('즐겨찾기 불러오기 에러: $e');
     }
@@ -43,39 +25,34 @@ class FavoriteService extends ChangeNotifier {
     required String destination,
     required String category,
   }) async {
-      final prefs = await SharedPreferences.getInstance();
+    final prefs = await SharedPreferences.getInstance();
     final googleId = prefs.getString('googleId');
     if (googleId == null) return;
-    final baseUrl = dotenv.env['BACKEND_URL'] ?? 'http://localhost:3000';
-    final uri = Uri.parse('$baseUrl/traffic/routes/favorites');
-    final payload = {
-      'googleId': googleId,
-      'origin': origin,
-      'destination': destination,
-      'category': category,
-    };
-    final resp = await http.post(uri,
-        headers: {'Content-Type': 'application/json'},
-        body: json.encode(payload));
-    if (resp.statusCode >= 200 && resp.statusCode < 300) {
+    try {
+      await FavoriteApiService(googleId: googleId).addFavorite(
+        origin: origin,
+        destination: destination,
+        category: category,
+        wakeUpTime: DateTime.now().toIso8601String(),
+      );
       await loadData();
-    } else {
-      throw Exception('즐겨찾기 추가 실패: ${resp.body}');
+    } catch (e) {
+      debugPrint('즐겨찾기 추가 에러: $e');
+      rethrow;
     }
   }
 
   Future<void> removeFavorite(String id) async {
-      final prefs = await SharedPreferences.getInstance();
+    final prefs = await SharedPreferences.getInstance();
     final googleId = prefs.getString('googleId');
     if (googleId == null) return;
-    final baseUrl = dotenv.env['BACKEND_URL'] ?? 'http://localhost:3000';
-    final uri = Uri.parse('$baseUrl/traffic/routes/favorites/$id?googleId=$googleId');
-    final resp = await http.delete(uri);
-    if (resp.statusCode >= 200 && resp.statusCode < 300) {
+    try {
+      await FavoriteApiService(googleId: googleId).removeFavorite(id);
       _favoriteList.removeWhere((f) => f.id == id);
       notifyListeners();
-      } else {
-      throw Exception('즐겨찾기 삭제 실패: ${resp.body}');
+    } catch (e) {
+      debugPrint('즐겨찾기 삭제 에러: $e');
+      rethrow;
     }
   }
