+++ lib/pages/auth/splash_screen.dart
