+++ lib/pages/home/home_page.dart
@@ -9,16 +9,16 @@ import 'package:permission_handler/permission_handler.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:wakelock_plus/wakelock_plus.dart';
 import 'package:http/http.dart' as http;
-import 'time_setting_page.dart';
-import 'arrivemappage.dart';
-import 'event_service.dart'; // EventService 불러오기
-import 'calendar_page.dart' hide EventService; // CalendarPage 불러오기
-import 'route_store.dart';
-import 'package:untitled4/models/route_response.dart';
-import 'favorite_service.dart';
-import 'favorite_management_page.dart';
-import 'models/favorite_route_model.dart';
-import 'services/alarm_api_service.dart';
+import '../time_setting/time_setting_page.dart';
+import '../search/search_page.dart';
+import '../../event_service.dart'; // EventService 불러오기
+import '../calendar/calendar_page.dart' hide EventService; // CalendarPage 불러오기
+import '../../route_store.dart';
+import '../../favorite_service.dart';
+import '../my/favorite_places_page.dart';
+import '../../services/alarm_api_service.dart';
+import '../../models/route_response.dart';
+import '../../models/favorite_route_model.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({Key? key}) : super(key: key);
@@ -547,237 +547,198 @@ class _HomePageState extends State<HomePage> {
         right: false,
         bottom: false,
         child: SingleChildScrollView(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 18.0, vertical: 16),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: [
-              // 검색창
-              GestureDetector(
-                onTap: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(builder: (context) => const ArriveMapPage()),
-                  );
-                },
-                child: Container(
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.circular(16),
-                    border: Border.all(color: Colors.black26, width: 1),
-                  ),
-                  child: AbsorbPointer(
-                    child: TextField(
-                      controller: searchController,
-                      decoration: const InputDecoration(
-                        hintText: '출발, 도착지 검색',
-                        prefixIcon: Icon(Icons.search, color: Colors.black54),
-                        border: InputBorder.none,
-                        contentPadding: EdgeInsets.symmetric(vertical: 14),
+          child: Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 18.0, vertical: 16),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              children: [
+                // 검색창
+                GestureDetector(
+                  onTap: () {
+                    Navigator.push(
+                      context,
+                      MaterialPageRoute(builder: (context) => const SearchPage()),
+                    );
+                  },
+                  child: Container(
+                    decoration: BoxDecoration(
+                      color: Colors.white,
+                      borderRadius: BorderRadius.circular(16),
+                      border: Border.all(color: Colors.black26, width: 1),
+                    ),
+                    child: AbsorbPointer(
+                      child: TextField(
+                        controller: searchController,
+                        decoration: const InputDecoration(
+                          hintText: '출발, 도착지 검색',
+                          prefixIcon: Icon(Icons.search, color: Colors.black54),
+                          border: InputBorder.none,
+                          contentPadding: EdgeInsets.symmetric(vertical: 14),
+                        ),
+                        style: const TextStyle(fontSize: 16),
                       ),
-                      style: const TextStyle(fontSize: 16),
                     ),
                   ),
                 ),
-              ),
-              SizedBox(height: verticalGap),
-              // 남은 준비 시간 카드 (통합된 버전) - 알람 상태 표시 추가
-              GestureDetector(
-                onTap: _goToTimeSettingPage,
-                child: Container(
-                  padding: const EdgeInsets.symmetric(vertical: 24, horizontal: 16),
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.circular(18),
-                    boxShadow: [
-                      BoxShadow(
-                        color: Colors.black12,
-                        blurRadius: 2,
-                        offset: Offset(0, 1),
-                      ),
-                    ],
-                    // 알람이 울릴 때 테두리 색상 변경
-                    border: _isAlarmRinging
-                        ? Border.all(color: Colors.red, width: 2)
-                        : null,
-                  ),
-                  child: Column(
-                    children: [
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: [
-                          const Text(
-                            '남은 준비 시간',
-                            style: TextStyle(
-                              fontSize: 16,
-                              color: Colors.black87,
-                              fontWeight: FontWeight.w500,
+                SizedBox(height: verticalGap),
+                // 남은 준비 시간 카드 (통합된 버전) - 알람 상태 표시 추가
+                GestureDetector(
+                  onTap: _goToTimeSettingPage,
+                  child: Container(
+                    padding: const EdgeInsets.symmetric(vertical: 24, horizontal: 16),
+                    decoration: BoxDecoration(
+                      color: Colors.white,
+                      borderRadius: BorderRadius.circular(18),
+                      boxShadow: [
+                        BoxShadow(
+                          color: Colors.black12,
+                          blurRadius: 2,
+                          offset: Offset(0, 1),
+                        ),
+                      ],
+                      // 알람이 울릴 때 테두리 색상 변경
+                      border: _isAlarmRinging
+                          ? Border.all(color: Colors.red, width: 2)
+                          : null,
+                    ),
+                    child: Column(
+                      children: [
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          children: [
+                            const Text(
+                              '남은 준비 시간',
+                              style: TextStyle(
+                                fontSize: 16,
+                                color: Colors.black87,
+                                fontWeight: FontWeight.w500,
+                              ),
                             ),
+                            if (_isAlarmRinging) ...[
+                              const SizedBox(width: 8),
+                              Icon(
+                                Icons.alarm,
+                                color: Colors.red,
+                                size: 20,
+                              ),
+                            ],
+                          ],
+                        ),
+                        const SizedBox(height: 8),
+                        Text(
+                          _formatDuration(remainingTime),
+                          style: TextStyle(
+                            fontSize: 40,
+                            fontWeight: FontWeight.bold,
+                            color: _isAlarmRinging ? Colors.red : Colors.black,
+                            letterSpacing: 2,
                           ),
-                          if (_isAlarmRinging) ...[
-                            const SizedBox(width: 8),
-                            Icon(
-                              Icons.alarm,
-                              color: Colors.red,
-                              size: 20,
+                        ),
+                        const SizedBox(height: 4),
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          children: [
+                            Text(
+                              _formatPrepTime(),
+                              style: const TextStyle(
+                                fontSize: 18,
+                                color: Colors.green,
+                                fontWeight: FontWeight.w500,
+                              ),
                             ),
                           ],
-                        ],
-                      ),
-                      const SizedBox(height: 8),
-                      Text(
-                        _formatDuration(remainingTime),
-                        style: TextStyle(
-                          fontSize: 40,
-                          fontWeight: FontWeight.bold,
-                          color: _isAlarmRinging ? Colors.red : Colors.black,
-                          letterSpacing: 2,
                         ),
-                      ),
-                      const SizedBox(height: 4),
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: [
-                          Text(
-                            _formatPrepTime(),
-                            style: const TextStyle(
-                              fontSize: 18,
-                              color: Colors.green,
-                              fontWeight: FontWeight.w500,
-                            ),
-                          ),
-                        ],
-                      ),
-                      const SizedBox(height: 16),
-                      // 도착시간과 알람예정시간 정보
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                        children: [
-                          // 설정 도착시간
-                          Column(
-                            children: [
-                              const Text(
-                                '설정 도착시간',
-                                style: TextStyle(
-                                  fontSize: 13,
-                                  color: Colors.black54,
-                                  fontWeight: FontWeight.w500,
-                                ),
-                              ),
-                              const SizedBox(height: 4),
-                              Text(
-                                _getArrivalTimeString(),
-                                style: const TextStyle(
-                                  fontSize: 16,
-                                  color: Colors.blue,
-                                  fontWeight: FontWeight.bold,
-                                ),
-                              ),
-                            ],
-                          ),
-                          // 구분선
-                          Container(
-                            height: 40,
-                            width: 1,
-                            color: Colors.black12,
-                          ),
-                          // 알람예정시간
-                          Column(
-                            children: [
-                              const Text(
-                                '알람예정시간',
-                                style: TextStyle(
-                                  fontSize: 13,
-                                  color: Colors.black54,
-                                  fontWeight: FontWeight.w500,
-                                ),
-                              ),
-                              const SizedBox(height: 4),
-                              Text(
-                                _getAlarmTimeString(),
-                                style: const TextStyle(
-                                  fontSize: 16,
-                                  color: Colors.orange,
-                                  fontWeight: FontWeight.bold,
-                                ),
-                              ),
-                            ],
-                          ),
-                        ],
-                      ),
-                      const SizedBox(height: 16),
-                      // 버튼 영역 - 상태에 따라 다른 버튼 표시
-                      if (!isAlarmScheduleActive && !isCountdownActive)
-                        ElevatedButton(
-                          onPressed: _startAlarmSchedule,
-                          style: ElevatedButton.styleFrom(
-                            backgroundColor: Colors.blue,
-                            foregroundColor: Colors.white,
-                            minimumSize: Size(160, 36),
-                            shape: RoundedRectangleBorder(
-                              borderRadius: BorderRadius.circular(18),
-                            ),
-                          ),
-                          child: const Text(
-                            '알람 시작',
-                            style: TextStyle(
-                              fontSize: 14,
-                              fontWeight: FontWeight.bold,
-                            ),
-                          ),
-                        )
-                      else if (isAlarmScheduleActive)
-                        Column(
+                        const SizedBox(height: 16),
+                        // 도착시간과 알람예정시간 정보
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                           children: [
-                            Row(
-                              mainAxisAlignment: MainAxisAlignment.center,
+                            // 설정 도착시간
+                            Column(
                               children: [
-                                Icon(Icons.access_time, color: Colors.blue, size: 16),
-                                const SizedBox(width: 4),
                                 const Text(
-                                  '알람예정시간 대기 중...',
+                                  '설정 도착시간',
                                   style: TextStyle(
-                                    fontSize: 14,
-                                    color: Colors.blue,
+                                    fontSize: 13,
+                                    color: Colors.black54,
                                     fontWeight: FontWeight.w500,
+                                  ),
+                                ),
+                                const SizedBox(height: 4),
+                                Text(
+                                  _getArrivalTimeString(),
+                                  style: const TextStyle(
+                                    fontSize: 16,
+                                    color: Colors.blue,
+                                    fontWeight: FontWeight.bold,
                                   ),
                                 ),
                               ],
                             ),
-                            const SizedBox(height: 8),
-                            ElevatedButton(
-                              onPressed: _stopAlarmSchedule,
-                              style: ElevatedButton.styleFrom(
-                                backgroundColor: Colors.red,
-                                foregroundColor: Colors.white,
-                                minimumSize: Size(160, 36),
-                                shape: RoundedRectangleBorder(
-                                  borderRadius: BorderRadius.circular(18),
+                            // 구분선
+                            Container(
+                              height: 40,
+                              width: 1,
+                              color: Colors.black12,
+                            ),
+                            // 알람예정시간
+                            Column(
+                              children: [
+                                const Text(
+                                  '알람예정시간',
+                                  style: TextStyle(
+                                    fontSize: 13,
+                                    color: Colors.black54,
+                                    fontWeight: FontWeight.w500,
+                                  ),
                                 ),
-                              ),
-                              child: const Text(
-                                '알람 취소',
-                                style: TextStyle(
-                                  fontSize: 14,
-                                  fontWeight: FontWeight.bold,
+                                const SizedBox(height: 4),
+                                Text(
+                                  _getAlarmTimeString(),
+                                  style: const TextStyle(
+                                    fontSize: 16,
+                                    color: Colors.orange,
+                                    fontWeight: FontWeight.bold,
+                                  ),
                                 ),
-                              ),
+                              ],
                             ),
                           ],
-                        )
-                      else if (isCountdownActive)
+                        ),
+                        const SizedBox(height: 16),
+                        // 버튼 영역 - 상태에 따라 다른 버튼 표시
+                        if (!isAlarmScheduleActive && !isCountdownActive)
+                          ElevatedButton(
+                            onPressed: _startAlarmSchedule,
+                            style: ElevatedButton.styleFrom(
+                              backgroundColor: Colors.blue,
+                              foregroundColor: Colors.white,
+                              minimumSize: Size(160, 36),
+                              shape: RoundedRectangleBorder(
+                                borderRadius: BorderRadius.circular(18),
+                              ),
+                            ),
+                            child: const Text(
+                              '알람 시작',
+                              style: TextStyle(
+                                fontSize: 14,
+                                fontWeight: FontWeight.bold,
+                              ),
+                            ),
+                          )
+                        else if (isAlarmScheduleActive)
                           Column(
                             children: [
                               Row(
                                 mainAxisAlignment: MainAxisAlignment.center,
                                 children: [
-                                  Icon(Icons.timer, color: Colors.green, size: 16),
+                                  Icon(Icons.access_time, color: Colors.blue, size: 16),
                                   const SizedBox(width: 4),
                                   const Text(
-                                    '준비시간 카운트다운 진행 중...',
+                                    '알람예정시간 대기 중...',
                                     style: TextStyle(
                                       fontSize: 14,
-                                      color: Colors.green,
+                                      color: Colors.blue,
                                       fontWeight: FontWeight.w500,
                                     ),
                                   ),
@@ -785,7 +746,7 @@ class _HomePageState extends State<HomePage> {
                               ),
                               const SizedBox(height: 8),
                               ElevatedButton(
-                                onPressed: _stopCountdown,
+                                onPressed: _stopAlarmSchedule,
                                 style: ElevatedButton.styleFrom(
                                   backgroundColor: Colors.red,
                                   foregroundColor: Colors.white,
@@ -795,7 +756,7 @@ class _HomePageState extends State<HomePage> {
                                   ),
                                 ),
                                 child: const Text(
-                                  '중지',
+                                  '알람 취소',
                                   style: TextStyle(
                                     fontSize: 14,
                                     fontWeight: FontWeight.bold,
@@ -803,196 +764,235 @@ class _HomePageState extends State<HomePage> {
                                 ),
                               ),
                             ],
-                          ),
-                    ],
+                          )
+                        else if (isCountdownActive)
+                            Column(
+                              children: [
+                                Row(
+                                  mainAxisAlignment: MainAxisAlignment.center,
+                                  children: [
+                                    Icon(Icons.timer, color: Colors.green, size: 16),
+                                    const SizedBox(width: 4),
+                                    const Text(
+                                      '준비시간 카운트다운 진행 중...',
+                                      style: TextStyle(
+                                        fontSize: 14,
+                                        color: Colors.green,
+                                        fontWeight: FontWeight.w500,
+                                      ),
+                                    ),
+                                  ],
+                                ),
+                                const SizedBox(height: 8),
+                                ElevatedButton(
+                                  onPressed: _stopCountdown,
+                                  style: ElevatedButton.styleFrom(
+                                    backgroundColor: Colors.red,
+                                    foregroundColor: Colors.white,
+                                    minimumSize: Size(160, 36),
+                                    shape: RoundedRectangleBorder(
+                                      borderRadius: BorderRadius.circular(18),
+                                    ),
+                                  ),
+                                  child: const Text(
+                                    '중지',
+                                    style: TextStyle(
+                                      fontSize: 14,
+                                      fontWeight: FontWeight.bold,
+                                    ),
+                                  ),
+                                ),
+                              ],
+                            ),
+                      ],
+                    ),
                   ),
                 ),
-              ),
-              SizedBox(height: verticalGap),
-              // 즐겨찾기 아이콘 섹션 추가
-              _buildFavoriteIcons(topFavorites),
-              SizedBox(height: verticalGap),
-              // 다음 경로 박스 (기존 교통수단 아이콘들을 대체)
-              GestureDetector(
-                onTap: _goToShortestRoutePage,
-                child: Container(
-                  padding: const EdgeInsets.symmetric(vertical: 20, horizontal: 24),
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.circular(16),
-                    border: Border.all(color: Colors.black26, width: 1),
-                    boxShadow: [
-                      BoxShadow(
-                        color: Colors.black12,
-                        blurRadius: 4,
-                        offset: const Offset(0, 2),
-                      ),
-                    ],
+                SizedBox(height: verticalGap),
+                // 즐겨찾기 아이콘 섹션 추가
+                _buildFavoriteIcons(topFavorites),
+                SizedBox(height: verticalGap),
+                // 다음 경로 박스 (기존 교통수단 아이콘들을 대체)
+                GestureDetector(
+                  onTap: _goToShortestRoutePage,
+                  child: Container(
+                    padding: const EdgeInsets.symmetric(vertical: 20, horizontal: 24),
+                    decoration: BoxDecoration(
+                      color: Colors.white,
+                      borderRadius: BorderRadius.circular(16),
+                      border: Border.all(color: Colors.black26, width: 1),
+                      boxShadow: [
+                        BoxShadow(
+                          color: Colors.black12,
+                          blurRadius: 4,
+                          offset: const Offset(0, 2),
+                        ),
+                      ],
+                    ),
+                    child: Row(
+                      children: [
+                        // 왼쪽 아이콘 영역
+                        Container(
+                          padding: const EdgeInsets.all(12),
+                          decoration: BoxDecoration(
+                            color: Colors.blue.shade50,
+                            borderRadius: BorderRadius.circular(12),
+                          ),
+                          child: Icon(
+                            Icons.directions,
+                            size: 32,
+                            color: Colors.blue,
+                          ),
+                        ),
+                        const SizedBox(width: 16),
+                        // 중앙 텍스트 영역
+                        Expanded(
+                          child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              const Text(
+                                '다음 경로',
+                                style: TextStyle(
+                                  fontSize: 20,
+                                  fontWeight: FontWeight.bold,
+                                  color: Colors.black87,
+                                ),
+                              ),
+                              const SizedBox(height: 4),
+                              const Text(
+                                '최적의 경로를 확인하세요',
+                                style: TextStyle(
+                                  fontSize: 14,
+                                  color: Colors.black54,
+                                ),
+                              ),
+                            ],
+                          ),
+                        ),
+                        // 오른쪽 화살표 아이콘
+                        Icon(
+                          Icons.arrow_forward_ios,
+                          size: 20,
+                          color: Colors.black54,
+                        ),
+                      ],
+                    ),
                   ),
-                  child: Row(
-                    children: [
-                      // 왼쪽 아이콘 영역
-                      Container(
-                        padding: const EdgeInsets.all(12),
-                        decoration: BoxDecoration(
-                          color: Colors.blue.shade50,
-                          borderRadius: BorderRadius.circular(12),
-                        ),
-                        child: Icon(
-                          Icons.directions,
-                          size: 32,
-                          color: Colors.blue,
-                        ),
-                      ),
-                      const SizedBox(width: 16),
-                      // 중앙 텍스트 영역
-                      Expanded(
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
+                ),
+                SizedBox(height: verticalGap),
+                // MM/DD 요일 + 오늘 일정 요약 카드
+                GestureDetector(
+                  onTap: _goToCalendarPage, // 캘린더 페이지로 이동
+                  child: Container(
+                    padding: const EdgeInsets.all(16),
+                    decoration: BoxDecoration(
+                      color: Colors.white,
+                      borderRadius: BorderRadius.circular(14),
+                      border: Border.all(color: Colors.black12, width: 1),
+                    ),
+                    child: Column(
+                      children: [
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
                           children: [
-                            const Text(
-                              '다음 경로',
-                              style: TextStyle(
-                                fontSize: 20,
-                                fontWeight: FontWeight.bold,
+                            Text(
+                              formattedDate,
+                              style: const TextStyle(
+                                fontSize: 22,
+                                fontWeight: FontWeight.w600,
                                 color: Colors.black87,
                               ),
                             ),
-                            const SizedBox(height: 4),
-                            const Text(
-                              '최적의 경로를 확인하세요',
-                              style: TextStyle(
-                                fontSize: 14,
-                                color: Colors.black54,
-                              ),
+                            const Icon(
+                              Icons.calendar_today,
+                              color: Colors.blue,
+                              size: 22,
                             ),
                           ],
                         ),
-                      ),
-                      // 오른쪽 화살표 아이콘
-                      Icon(
-                        Icons.arrow_forward_ios,
-                        size: 20,
-                        color: Colors.black54,
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-              SizedBox(height: verticalGap),
-              // MM/DD 요일 + 오늘 일정 요약 카드
-              GestureDetector(
-                onTap: _goToCalendarPage, // 캘린더 페이지로 이동
-                child: Container(
-                  padding: const EdgeInsets.all(16),
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.circular(14),
-                    border: Border.all(color: Colors.black12, width: 1),
-                  ),
-                  child: Column(
-                    children: [
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [
-                          Text(
-                            formattedDate,
-                            style: const TextStyle(
-                              fontSize: 22,
-                              fontWeight: FontWeight.w600,
-                              color: Colors.black87,
-                            ),
+                        const SizedBox(height: 12),
+                        const Divider(height: 1, color: Colors.black12),
+                        const SizedBox(height: 12),
+                        const Text(
+                          '오늘 일정',
+                          style: TextStyle(
+                            fontSize: 18,
+                            fontWeight: FontWeight.w500,
+                            color: Colors.black87,
                           ),
-                          const Icon(
-                            Icons.calendar_today,
-                            color: Colors.blue,
-                            size: 22,
-                          ),
-                        ],
-                      ),
-                      const SizedBox(height: 12),
-                      const Divider(height: 1, color: Colors.black12),
-                      const SizedBox(height: 12),
-                      const Text(
-                        '오늘 일정',
-                        style: TextStyle(
-                          fontSize: 18,
-                          fontWeight: FontWeight.w500,
-                          color: Colors.black87,
+                          textAlign: TextAlign.center,
                         ),
-                        textAlign: TextAlign.center,
-                      ),
-                      const SizedBox(height: 8),
-                      if (todayEvents.isEmpty)
-                        const Padding(
-                          padding: EdgeInsets.symmetric(vertical: 12.0),
-                          child: Text(
-                            '오늘 예정된 일정이 없습니다',
-                            style: TextStyle(
-                              fontSize: 15,
-                              color: Colors.black54,
+                        const SizedBox(height: 8),
+                        if (todayEvents.isEmpty)
+                          const Padding(
+                            padding: EdgeInsets.symmetric(vertical: 12.0),
+                            child: Text(
+                              '오늘 예정된 일정이 없습니다',
+                              style: TextStyle(
+                                fontSize: 15,
+                                color: Colors.black54,
+                              ),
+                              textAlign: TextAlign.center,
                             ),
-                            textAlign: TextAlign.center,
-                          ),
-                        )
-                      else
-                        ListView.builder(
-                          shrinkWrap: true,
-                          physics: const NeverScrollableScrollPhysics(),
-                          itemCount: todayEvents.length > 3 ? 3 : todayEvents.length, // 최대 3개만 표시
-                          itemBuilder: (context, index) {
-                            final event = todayEvents[index]; // 이벤트 객체를 미리 가져오기
-                            return Padding(
-                              padding: const EdgeInsets.symmetric(vertical: 4.0),
-                              child: Container(
-                                padding: const EdgeInsets.all(12),
-                                decoration: BoxDecoration(
-                                  color: Colors.blue.shade50,
-                                  borderRadius: BorderRadius.circular(8),
-                                  border: Border.all(color: Colors.blue.shade200),
-              ),
-              child: Row(
-                children: [
-                  Icon(
-                    Icons.event,
-                    color: Colors.blue,
-                    size: 16,
-                  ),
-                  const SizedBox(width: 8),
-                  Expanded(
-                    child: Text(
-                      event.title ?? event.content ?? '제목 없음', // title이 있으면 title, 없으면 content 사용
-                      style: const TextStyle(
-                        fontSize: 14,
-                        fontWeight: FontWeight.w500,
-                        color: Colors.black87,
-                      ),
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                  ),
-                  if (event.time != null && event.time!.isNotEmpty) // time 속성 사용
-                    Text(
-                      event.time!,
-                      style: const TextStyle(
-                        fontSize: 12,
-                        color: Colors.black54,
-                      ),
-                    ),
-                ],
-              ),
-            ),
-          );
-        },
+                          )
+                        else
+                          ListView.builder(
+                            shrinkWrap: true,
+                            physics: const NeverScrollableScrollPhysics(),
+                            itemCount: todayEvents.length > 3 ? 3 : todayEvents.length, // 최대 3개만 표시
+                            itemBuilder: (context, index) {
+                              final event = todayEvents[index]; // 이벤트 객체를 미리 가져오기
+                              return Padding(
+                                padding: const EdgeInsets.symmetric(vertical: 4.0),
+                                child: Container(
+                                  padding: const EdgeInsets.all(12),
+                                  decoration: BoxDecoration(
+                                    color: Colors.blue.shade50,
+                                    borderRadius: BorderRadius.circular(8),
+                                    border: Border.all(color: Colors.blue.shade200),
+                                  ),
+                                  child: Row(
+                                    children: [
+                                      Icon(
+                                        Icons.event,
+                                        color: Colors.blue,
+                                        size: 16,
+                                      ),
+                                      const SizedBox(width: 8),
+                                      Expanded(
+                                        child: Text(
+                                          event.title ?? event.content ?? '제목 없음', // title이 있으면 title, 없으면 content 사용
+                                          style: const TextStyle(
+                                            fontSize: 14,
+                                            fontWeight: FontWeight.w500,
+                                            color: Colors.black87,
+                                          ),
+                                          overflow: TextOverflow.ellipsis,
+                                        ),
+                                      ),
+                                      if (event.time != null && event.time!.isNotEmpty) // time 속성 사용
+                                        Text(
+                                          event.time!,
+                                          style: const TextStyle(
+                                            fontSize: 12,
+                                            color: Colors.black54,
+                                          ),
+                                        ),
+                                    ],
+                                  ),
+                                ),
+                              );
+                            },
                           ),
                       ],
                     ),
+                  ),
+                ),
+              ],
+            ),
           ),
         ),
-    ],
-    ),
-    ),
-    ),
-    ),
+      ),
     );
   }
 
@@ -1029,7 +1029,7 @@ class _HomePageState extends State<HomePage> {
                   Navigator.push(
                     context,
                     MaterialPageRoute(
-                      builder: (context) => const FavoriteManagementPage(),
+                      builder: (context) => const FavoritePlacesPage(),
                     ),
                   ).then((_) => _favoriteService.loadData());
                 },
@@ -1073,7 +1073,7 @@ class _HomePageState extends State<HomePage> {
         Navigator.push(
           context,
           MaterialPageRoute(
-            builder: (context) => ArriveMapPage(
+            builder: (context) => SearchPage(
               initialDestination: favorite.destination,
               initialDestinationAddress: favorite.destination,
             ),
@@ -1127,7 +1127,7 @@ class _HomePageState extends State<HomePage> {
         Navigator.push(
           context,
           MaterialPageRoute(
-            builder: (context) => const FavoriteManagementPage(),
+            builder: (context) => const FavoritePlacesPage(),
           ),
         ).then((_) => _favoriteService.loadData());
       },
